uq_pinn_mfl_research/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uq-pinn-mfl-research"
version = "0.1.0"
description = "Reviewer-constrained UQ-PINN + denoising + MFL research pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "scikit-learn>=1.5",
  "torch>=2.5",
]

[project.scripts]
uq-pinn-mfl = "uq_pinn_mfl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
